requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pandas==2.0.3
openpyxl==3.1.2
Pillow==12.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
cloudinary==1.36.0
requests==2.31.0